refactor: consolidate js and py runner modules into runner
