+++ src/main.rs
@@ -14,10 +14,9 @@ mod experiments;
 mod functions;
 mod http;
 mod init;
-mod js_runner;
 mod projects;
 mod prompts;
-mod python_runner;
+mod runner;
 mod scorers;
 mod self_update;
 mod setup;
