+++ src/runner/py.rs
@@ -1,4 +1,4 @@
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
 pub fn resolve_python_interpreter(
     explicit: Option<&str>,
@@ -42,34 +42,7 @@ fn find_virtual_env_python() -> Option<PathBuf> {
 }
 
 pub fn find_binary_in_path(candidates: &[&str]) -> Option<PathBuf> {
-    let paths = std::env::var_os("PATH")?;
-    for dir in std::env::split_paths(&paths) {
-        for candidate in candidates {
-            let path = dir.join(candidate);
-            if path.is_file() {
-                return Some(path);
-            }
-            if cfg!(windows) {
-                let exe = with_windows_extensions(&path);
-                for candidate_path in exe {
-                    if candidate_path.is_file() {
-                        return Some(candidate_path);
-                    }
-                }
-            }
-        }
-    }
-    None
-}
-
-#[cfg(windows)]
-fn with_windows_extensions(path: &Path) -> [PathBuf; 2] {
-    [path.with_extension("exe"), path.with_extension("cmd")]
-}
-
-#[cfg(not(windows))]
-fn with_windows_extensions(_path: &Path) -> [PathBuf; 0] {
-    []
+    super::find_binary_in_path(candidates)
 }
 
 #[cfg(test)]
